Move Gdx Pixmap code to native class name
